Refactored folder structure, fixed "username" typo, got rid of design error: double or divider, removed unused imports
