+++ lib/pages/take_action.dart
@@ -1,4 +1,3 @@
-import 'package:call_black_line/create_account.dart';
 import 'package:call_black_line/widgets/checkbox_text.dart';
 import 'package:call_black_line/widgets/orange_button.dart';
 import 'package:flutter/material.dart';
